config(git-hooks): add condition to run bash

+++ .githooks/main.sh
@@ -1,5 +1,7 @@
 #! /usr/bin/bash
-npm run prepare &
-wait
-cp .githooks/commit-msg .husky/
-cp .githooks/pre-commit .husky/
+if [ ! -e "./.husky" ]; then
+  npx husky &
+  wait
+  cp .githooks/commit-msg .husky/
+  cp .githooks/pre-commit .husky/
+fi
